Refactor top bar into its own module
